requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Plugwise specific (if available)
# plugwise-api==0.1.0  # Uncomment when you find the right library

# Optional: For advanced features
# aiohttp==3.9.1  # For async HTTP requests
# asyncio-mqtt==0.16.1  # For MQTT integration
# influxdb-client==1.38.0  # For InfluxDB integration 